requirements.txt
aiohttp==3.7.4
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==3.0.1
attrs==23.1.0
billiard==4.2.0
boto3==1.28.54
botocore==1.31.54
bson==0.5.10
celery==5.3.4
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-redis==5.4.0
django-relations==0.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt
drf-yasg==1.21.7
fastapi==0.103.2
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.4
MarkupSafe==2.1.3
minio==7.1.17
multidict==6.1.0
openapi-codec==1.3.2
packaging==24.2
pillow==11.0.0
prompt-toolkit==3.0.41
propcache==0.2.1
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.32.3
rich==10.14.0
s3transfer==0.6.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
starlette==0.27.0
typing_extensions
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
vine==5.1.0
wcwidth==0.2.12
websockets==9.1
yarl==1.18.3
psycopg2-binary
